+++ src/itemshortcut.cpp
@@ -74,9 +74,8 @@ void ItemShortcut::load(bool oldConfig)
     }
     for (int i = 0; i < SHORTCUT_ITEMS; i++)
     {
-        int itemId = static_cast<int>(cfg->getValue(name + toString(i), -1));
-        int itemColor = static_cast<int>(
-            cfg->getValue(color + toString(i), 1));
+        int itemId = cfg->getValue(name + toString(i), -1);
+        int itemColor = cfg->getValue(color + toString(i), 1);
 
         mItems[i] = itemId;
         mItemColors[i] = itemColor;
